requirements.txt
Flask
scikit-learn
numpy
pandas
werkzeug
gunicorn
python-magic-bin